Player can now be steered by arrow keys (no gravity)
JBox2d operates in meters, which without scaling means 1px == 1m, which
causes everything to appear very slow.  Adjusting to 50px per 1m results
in a much more realistic looking simulation.

+++ PixelBlaster/src/com/omnicrola/pixelblaster/map/IMapTile.java
@@ -4,6 +4,6 @@ import com.omnicrola.pixelblaster.graphics.IGraphicsWrapper;
 
 public interface IMapTile {
 
-	void render(int x, int y, IGraphicsWrapper graphics);
+	void render(float x, float y, IGraphicsWrapper graphics);
 
 }
